refactor: extract ManualCreatorCard components for improved readability and maintainability
